Restructure and optimize code
`Converter` class is now split into a `from_map16` and `to_map16` class, each with just a single public `convert` method. 

`~` is now used instead of `130 { 004 4 ---  004 4 ---  004 4 ---  004 4 --- }` to specify Lunar Magic's empty tile. This cuts the size of the output down from ~4MB to ~1.1MB for a vanilla ROM's map16 (if all tiles are in use the size will of course be back to ~4MB). The conversion also runs slightly faster thanks to this change.