Prevent Lunar Magic bug
Lunar Magic seemingly has a bug that will overwrite tileset specific page 2 of tileset E's map16 if you have the tileset specific page 2 setting enabled. I don't know why, but in this case the page 2 in the global FG pages (pointer 0 in the map16 file's offset + size table) overwrites tileset E's page 2 with page 2 of the tileset that was in use when the full map16 was exported (whatever tileset is in use in the level at the time, if exported via command line it's tileset 0).

To prevent this bug, when converting from txt -> .map16, tileset E's tileset specific page 2 is copied to global page 2 for now (idk if fusoya wants to fix this or if it's even a bug).